Complete foundations/06-cascade-fix
My solutions were not all mentioned exactly in the solutions page.
However, I did get the desired outcome. I reorganized the blocks so
that the blocks starting with the div selector are at the top because
all the other blocks start with more specific selector types, even if
the overall specificity may be less. I'm not sure if page order or
specificity order should be the preferred way of organizing the sheet.